feat: query using debug element, and wrap debug element in TestElement
A debugElement is more powerful than a nativeElement, since it allows querying by other means, and wraps a native element anyway.
So, this allows doing more complex queries, such as

```
   this.element('.section')
       .elements('.chapter')[1]
       .debugElement.query(By.directive(ImageComponent))
       .componentInstance
```

for example, to get the component instance of the first image of teh second chapter of the first section.

The tester itself now also has a wrapped element. This means that if a utility function needs a TestElement as argument, and we need to pass the element of the component under test itself to this function, we now can.